#29
When parameters in url is applied to map, the layer tree also unfolds.
(This functionality is perhaps better suited in the core or in
LayerTree? For not-to-break-anything purposes I’ll keep it here for now)

Introduced parameter ’preset’, who needs to be included in the
config/layers/params-string. When in url, the corresponding radio
button is checked on load.